fix of field type

+++ apc/storefront/shoppingcart/models.py
@@ -19,7 +19,7 @@ class Item(models.Model):
 
 class MenuItem(models.Model):
     name = models.CharField(max_length= 200)
-    price = models.FloatField(max_digits=10, decimal_places=2)
+    price = models.DecimalField(max_digits=4, decimal_places=2)
     description = models.CharField(max_length= 300)
     image = models.ImageField(upload_to='menu_images/')
